LayoutTests/platform/win/editing/selection/click-in-padding-with-multiple-line-boxes-expected.txt

Click above/beneath the word FOO. Notice how the cursor goes in the right place (after the letter F).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Selection is at node: [object Text] offset: 1
PASS Selection is at node: [object Text] offset: 1
PASS successfullyParsed is true

TEST COMPLETE
